+++ src/main/java/com/clinicmaster/clinic/domain/PayMent.java
@@ -1,14 +1,9 @@
 package com.clinicmaster.clinic.domain;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -16,6 +11,6 @@ import java.util.List;
 @Setter
 public class PayMent {
 
-    private int totalMoney;
+    private BigDecimal totalMoney;
     private List<MedicinePayMent> payMents;
 }
